Project 5/requirements.txt
SpeechRecognition==3.14.3
pyttsx3==2.99
requests==2.32.4
openai==1.10.0
gTTS==2.5.4
pygame==2.6.1
PyAudio==0.2.14
python-dotenv==1.0.1
